+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.21.1 (unreleased)
 
+- Added bitwise methods to `Expr`
 - Added `drop_nans` method to `DataFrame` and `LazyFrame`
 - Fixed `subset` option for `drop_nulls` method